backend/requirements.txt
# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# ============================================================================
# DATABASE & CACHING
# ============================================================================
# PostgreSQL
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0  # Async PostgreSQL driver
psycopg2-binary==2.9.9  # Sync PostgreSQL driver (backup)

# Redis
redis[hiredis]==5.0.1  # hiredis for performance

# SQLModel (ORM)
sqlmodel==0.0.14

# Database Migrations
alembic==1.13.1

# ============================================================================
# STORAGE
# ============================================================================
# S3/MinIO
aioboto3==12.3.0
boto3==1.34.10

# ============================================================================
# AI & MACHINE LEARNING
# ============================================================================
openai==1.6.1  # OpenAI API (GPT-4, GPT-4o, GPT-4o-mini)
anthropic==0.25.0  # Claude API (for monitoring auto-fix)
pillow==11.1.0  # Image processing (updated for HEIC support)
pillow-heif==1.1.1  # HEIC support (requires libheif 1.19+)
imagehash==4.3.1  # Perceptual hashing for duplicate detection

# ============================================================================
# AUTOMATION
# ============================================================================
playwright==1.40.0  # Browser automation
httpx==0.25.2  # Async HTTP client
requests==2.31.0  # Sync HTTP client (for Telegram notifications)

# ============================================================================
# MONITORING & OBSERVABILITY
# ============================================================================
sentry-sdk[fastapi]==1.39.1  # Error tracking
prometheus-client==0.19.0  # Metrics

# ============================================================================
# SECURITY
# ============================================================================
cryptography==41.0.7  # Encryption (AES-256)
argon2-cffi==23.1.0  # Password hashing
pyjwt==2.8.0  # JWT tokens
python-jose[cryptography]==3.3.0  # JWT utilities
itsdangerous==2.2.0  # Secure signing and serialization
slowapi==0.1.9  # Rate limiting

# ============================================================================
# PAYMENTS
# ============================================================================
stripe==7.8.0  # Stripe payments

# ============================================================================
# EMAIL
# ============================================================================
aiosmtplib==3.0.1  # Async SMTP
jinja2==3.1.2  # Email templates

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0  # Environment variables
loguru==0.7.2  # Logging
pydantic==2.5.2  # Data validation
pydantic-settings==2.1.0  # Settings management
email-validator==2.1.0  # Email validation
aiofiles==23.2.1  # Async file I/O
psutil==5.9.6  # System metrics
apscheduler==3.10.4  # Background jobs
python-magic==0.4.27  # File type detection
filetype==1.2.0  # File type detection
rapidfuzz==3.5.2  # Fuzzy string matching

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1  # Code formatting
ruff==0.1.8  # Linting
colorama==0.4.6  # Colored terminal output (for setup tests)
